fix(PDS-82): moved props definitions to specific files
