+++ src/components/Typography/HX/H3/index.tsx
@@ -18,7 +18,8 @@
 
 import { ReactElement } from 'react'
 
-import { HX, HXProps } from '../HX'
+import { HX } from '../HX'
+import { HXProps } from '../HX.props'
 
 export const H3 = (props: HXProps): ReactElement => {
   return <HX {...props} level={3} role={'h3'} />
