+++ src/components/Typography/HX/H4/index.tsx
@@ -18,7 +18,8 @@
 
 import { ReactElement } from 'react'
 
-import { HX, HXProps } from '../HX'
+import { HX } from '../HX'
+import { HXProps } from '../HX.props'
 
 export const H4 = (props: HXProps): ReactElement => {
   return <HX {...props} level={4} role={'h4'} />
